feat: Add actions (execute function from UI)
You are now able to define actions, these are UI elements that, instead of navigating to another window, execute a function you have defined.
