+++ new/Application.py
@@ -1,4 +1,3 @@
-import os
 from Window import Window
 
 class Application():
@@ -11,14 +10,8 @@ class Application():
             if window.get_name() == name:
                 return index
         return -1
-
-    def __clear_console(self) -> None:
-        if os.name == 'nt': os.system('cls')
-        else: os.system('clear')
         
     def run(self) -> None:
-
-        self.__clear_console()
 
         # Returns the option the user has selected in the active window
         user_choice = self.__activeWindow.run()
