+++ new/Window.py
@@ -1,5 +1,7 @@
 import sys
+import os
 from elements.Navigation import Navigation
+from elements.Action import Action
 
 class Window():
 
@@ -17,6 +19,10 @@ class Window():
 
     def __str__(self) -> str:
         return self.__name
+
+    def __clear_console(self) -> None:
+        if os.name == 'nt': os.system('cls')
+        else: os.system('clear')
 
     def get_name(self) -> str:
         return self.__name
@@ -37,11 +43,20 @@ class Window():
 
     def run(self) -> dict:
 
+        self.__clear_console()
+
         # Render all elements that belong to this window
         for element in self.__elements:
             element.display(self.__color_configuration, self.__width)
 
         user_input = self.getch()
+
+        # Execute the action inside the window
+        for element in [x for x in self.__elements if type(x) == Action]:
+            if element.get_abbreviation() == user_input:
+                self.__clear_console()
+                element.execute()
+                self.run()
 
         # Retrieve what the user input leads to (abbreviation to name
         for element in [x for x in self.__elements if type(x) == Navigation]:
