+++ new/configuration/Actions.py
@@ -1,0 +1,5 @@
+from configuration.Settings import Settings
+
+def generate_image() -> None:
+    print(Settings.test_message)
+    input("Wait...")
