+++ new/configuration/Settings.py
@@ -1,0 +1,5 @@
+class Settings():
+
+    image_output_directory = "images"
+    test_message = "Hello World"
+    openai_status = "Running"
