+++ new/elements/Action.py
@@ -1,10 +1,9 @@
-from elements.Element import Element
+from elements.Navigation import Navigation
 
-class Action(Element):
-    def __init__(self) -> None:
-        super().__init__()
+class Action(Navigation):
+    def __init__(self, abbreviation: str, name: str, variable: str = None, action = None) -> None:
+        super().__init__(abbreviation, name, variable)
+        self.__action = action
 
-    def display(self, color_configuration: dict, width: int) -> None:
-        # TBA
-        # ...
-        pass
+    def execute(self) -> any:
+        return self.__action()
